update code PIM Tool

+++ on-boarding-exercise/src/main/java/vn/elca/training/repository/custom/ProjectRespositoryCustom.java
@@ -6,5 +6,5 @@ import vn.elca.training.model.entity.Project;
 import java.util.List;
 
 public interface ProjectRespositoryCustom {
-    List<Project> findProjectById(Long id);
+    List<Project> findProjectById(Integer id);
 }
