+++ on-boarding-exercise/src/main/java/vn/elca/training/service/ProjectService.java
@@ -12,8 +12,8 @@ import vn.elca.training.model.entity.StatusProject;
 public interface ProjectService {
     List<Project> findAll();
     List<Project> findByCriteria(String name, StatusProject status);
-    List<Project> findById(Long id);
+    List<Project> findById(Integer id);
     void createNewProject(Project project);
     void deleteProject(Integer id);
-    void updateProject(Long id, Project project);
+    void updateProject(Integer id, Project project);
 }
